+++ android/src/main/java/com/reactnativekeyboardcontroller/managers/OverKeyboardViewManagerImpl.kt
@@ -2,7 +2,7 @@ package com.reactnativekeyboardcontroller.managers
 
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ThemedReactContext
-import com.reactnativekeyboardcontroller.views.OverKeyboardHostView
+import com.reactnativekeyboardcontroller.views.overlay.OverKeyboardHostView
 
 @Suppress("detekt:UnusedPrivateProperty")
 class OverKeyboardViewManagerImpl(mReactContext: ReactApplicationContext) {
